feat: added method to create payment in service.

+++ ms-payments/src/main/java/uol/compass/payments/dto/request/CalculateRequest.java
@@ -7,6 +7,6 @@ import lombok.*;
 public class CalculateRequest {
 
     private Long categoryId;
-    private Long value;
+    private Integer value;
 
 }
